% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_result.R
\name{fetch_result}
\alias{fetch_result}
\title{Fetches the result from a given URL with optional query parameters and
authentication.}
\usage{
fetch_result(url, .query_params = list(), .auth = NULL, return_request = FALSE)
}
\arguments{
\item{url}{The API endpoint to fetch the result from.}

\item{.query_params}{Optional query parameters to include in the request.}

\item{.auth}{Optional path to authentication cookie to use.}

\item{return_request}{If TRUE, the function returns the request object
instead of the response.}
}
\value{
The response object or the request object if return_request is TRUE.
}
\description{
Fetches the result from a given URL with optional query parameters and
authentication.
}
\examples{
\dontrun{
# Fetch result from a URL
fetch_result("https://uat-db.vespawatch.be/observations/30608")

# Fetch result with query parameters
fetch_result("https://uat-db.vespawatch.be/observations/",
             .query_params = list(municipality_id = 215))
# Return request object instead of response
fetch_result("https://uat-db.vespawatch.be/observations/30608",
            return_request = TRUE)
}
}
