% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_field_definitions.R
\name{get_field_definitions}
\alias{get_field_definitions}
\title{Get field definitions}
\usage{
get_field_definitions(domain = c("uat", "production"), check_required = TRUE)
}
\arguments{
\item{domain}{The domain to query. Either "uat" for the testing instance, or
"prod" for the production instance.}

\item{check_required}{Logical indicating whether to check if any fields are
required. If no required fields are in the definition a warning is
returned.}
}
\value{
A tibble with the field definitions.
}
\description{
This function retrieves the field definitions for the vespa-db Observations
API schema
}
\examples{
\dontrun{
# complete field definitions
get_field_definitions("prod")
get_field_definitions()
}

}
