% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observation_parameters.R
\name{get_observation_parameters}
\alias{get_observation_parameters}
\title{Show an overview of the available parameters for \code{\link[=get_vespadb_obs]{get_vespadb_obs()}}}
\usage{
get_observation_parameters(domain = c("production", "uat"), ...)
}
\arguments{
\item{domain}{The domain to login to. Default is "uat" for the UAT
environment.}

\item{...}{Additional parameters to internal methods.}
}
\value{
A tibble with the parameters that can be used in the
\code{\link[=get_vespadb_obs]{get_vespadb_obs()}} function.
}
\description{
This function retrieves the parameters that can be used in the
\code{\link[=get_vespadb_obs]{get_vespadb_obs()}} function. It will return a tibble with the parameters
that can be used in the query.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_observation_parameters()

\dontshow{\}) # examplesIf}
}
