% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vespadb_export_s3.R
\name{get_vespadb_export_s3}
\alias{get_vespadb_export_s3}
\title{This fn also returns invisible obs}
\usage{
get_vespadb_export_s3(domain = c("uat", "production"))
}
\arguments{
\item{domain}{The domain to query. Either "uat" for the testing instance, or
"prod" for the production instance.}
}
\description{
This fn also returns invisible obs
}
