% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vespadb_obs.R
\name{get_vespadb_obs}
\alias{get_vespadb_obs}
\title{Query the vespa-db api to get observations}
\usage{
get_vespadb_obs(..., auth = NULL, domain = c("uat", "production"))
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Query parameters according to
the observation model.}

\item{auth}{Character. Path to the authentication cookie to use. By default a
request without authentication is placed (public data only). The
authentication cookie can be created using \code{\link[=login_vespadb]{login_vespadb()}} which returns
the path to the cookie.}

\item{domain}{Character. The domain to login to. Default is "uat" for the UAT
environment.}
}
\value{
A tibble with the observations from the VespaDB API.
}
\description{
Extra query parameters are supported according to the Observation model. More
information on the allowed parameters can be fetched via
\code{\link[=get_field_definitions]{get_field_definitions()}}.
}
\details{
This function already handles paging, so supplying a page as a parameter is
not allowed. For a more low level function to fetch results see
\code{\link[=fetch_result]{fetch_result()}}.
}
\examples{
\dontrun{
# All public records
get_vespadb_obs()

# Public records from Ghent
get_vespadb_obs(municipality_id = 215)

# Public and private information from Ghent, only ANB plots
get_vespadb_obs(
  municipality_id = 215,
  anb = "true",
  auth = login_vespadb()
)
}

}
