% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_vespadb.R
\name{login_vespadb}
\alias{login_vespadb}
\title{Login to VespaDB API to fetch fields that are not publicly available}
\usage{
login_vespadb(username, password, domain = c("production", "uat"))
}
\arguments{
\item{username}{The username to login with, if not provided it will
be read from the environment variable \code{VESPADB_USER}. If that is not set,
you will be prompted to enter their username.}

\item{password}{The password to login with, if not provided it will
be read from the environment variable \code{VESPADB_PWD}. If that is not set,
you will be prompted to enter their password.}

\item{domain}{The domain to login to. Default is "production" for the UAT
environment.}
}
\value{
The path to an authentication cookie.
}
\description{
This function allows you to login to the API and returns the required
csrftoken needed to fetch information that is not publicly available.
}
\examples{
\dontrun{
get_vespadb_obs(
  municipality = 215,
  auth = login_vespadb(
    "username",
    "password"
  )
)
}
}
