% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\name{municipalities}
\alias{municipalities}
\title{Get the municipality ids.}
\usage{
municipalities(municipalities = NULL, domain = c("production", "uat"))
}
\arguments{
\item{municipalities}{(Optional) A character vector of municipality names. If
this is not provided, all municipalities will be returned as a tibble with
their ids.}
}
\value{
Either a numeric vector of ids, or if no argument is passed, a tibble
with all municipalities and their id.
}
\description{
This function either returns a complete overview of all municipality ids. Or
if one or more municipalities are passed, the id's for these municipalities.
}
\examples{
municipalities("Gent")
municipalities(c("Gent", "Kortrijk"))
municipalities()
}
