% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespadb-package.R
\docType{package}
\name{vespadb-package}
\alias{vespadb}
\alias{vespadb-package}
\title{vespadb: Read Data from the vespa-db Database}
\description{
Read data from the vespa-db database hosted at https://nesten.vespawatch.be.
}
\author{
\strong{Maintainer}: Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID}) (Research Institute for Nature and Forest (INBO))

Authors:
\itemize{
  \item Sander Devisscher \email{sander.devisscher@inbo.be} (\href{https://orcid.org/0000-0003-2015-5731}{ORCID}) (Research Institute for Nature and Forest (INBO))
}

}
\keyword{internal}
