% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vespadb_export_s3.R
\name{get_vespadb_export_s3}
\alias{get_vespadb_export_s3}
\title{This fn also returns invisible obs}
\usage{
get_vespadb_export_s3(domain = c("uat", "production"))
}
\description{
This fn also returns invisible obs
}
