% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_vespadb.R
\name{login_vespadb}
\alias{login_vespadb}
\title{Login to VespaDB API to fetch fields that are not publicly available}
\usage{
login_vespadb(username, password, domain = c("uat", "production"))
}
\arguments{
\item{username}{The username to login with}

\item{password}{The password to login with}

\item{domain}{The domain to login to. Default is "uat" for the UAT
environment.}

\item{verbose}{Whether to print the response from the API. Default is FALSE.}
}
\value{
The path to an authentication cookie.
}
\description{
This function allows you to login to the API and returns the required
csrftoken needed to fetch information that is not publicly available.
}
\examples{
\dontrun{
get_vespadb_obs(municipality = 215,
                          auth = login_vespadb("username",
                                               "password"))
}
}
